requirements.txt
google-genai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
rank-bm25>=0.2.2
streamlit>=1.35.0
Pillow>=10.0.0
python-dotenv>=1.0.0
